Set timezone based on TZ environment variable
Setting the TZ environment variable won't affect the timezone
of the processes started by the supervisor since it trims the
environment passed to the started child processes.

The proper way to set the timezone in Debian is to adjust the
/etc/localtime symlink and reconfigure the tzdata package. We
do this in an early setup script to not get a mix of timezones
in e.g. logs.

Fixes #123